+++ auto/attack.js
@@ -40,9 +40,9 @@ async function updateAttack(channel) {
 
             //if not in db and active, post and db_save new chat
             if (!chat && attack.status === 'active') {
-                const content = generate_defence_message(api); // create message content
+                const content = generate_attack_message(attack); // create message content
                 const message = await channel.send(content); // post message to discord (returns message object)
-                const event = await db_SaveEvent(api.event_id, message.id); // save event with --linked messageId --- to database
+                const event = await db_SaveEvent(attack.event_id, message.id); // save event with --linked messageId --- to database
                 log.info(
                     chalk.cyan('attack.js') +
                         chalk.white(' updateAttack() created event(') +
@@ -59,7 +59,7 @@ async function updateAttack(channel) {
             if (chat && chat.active) {
                 deleted = chat.event_id; //precaution if the message has been deleted, variable will be used in the catch block below
                 const message = await channel.messages.fetch(chat.message_id); // fetch message from discord by id -> this will error if the message is deleted
-                const content = generate_defence_message(attack, chat); // create message content
+                const content = generate_attack_message(attack, chat); // create message content
                 const update = await message.edit(content); // update the message with new content
                 const event = await db_updateEvent(attack); // update the database record
                 log.info(