+++ prisma/migrations/20241227135323_finished_column/migration.sql
@@ -1,5 +1,0 @@
--- AlterTable
-ALTER TABLE `Event` ADD COLUMN `finished` BOOLEAN NOT NULL DEFAULT false;
-
--- AlterTable
-ALTER TABLE `Post` ADD COLUMN `finished` BOOLEAN NOT NULL DEFAULT false;