+++ prisma/migrations/20250104041137_asd/migration.sql
@@ -1,2 +1,0 @@
--- AlterTable
-ALTER TABLE `Defend` ADD COLUMN `active` BOOLEAN NOT NULL DEFAULT true;